added veritable onbarding and deprovisioning workflow

+++ app/temporalworkflows/common/configmap.py
@@ -8,10 +8,10 @@ class ConfigMap:
     ConfigMap class for creating and updating ConfigMap
     """
 
-    def __init__(self, tenant: str):
-        self.tenant: str = tenant
+    def __init__(self, namespace: str):
+        self.namespace: str = namespace
 
-    def apply_config(self, name: str, data: dict):
+    def create_or_replace_config(self, name: str, data: dict):
         """
         Apply the ConfigMap
         """
@@ -20,6 +20,27 @@ class ConfigMap:
         try:
             with client.ApiClient() as api_client:
                 api_instance = client.CoreV1Api(api_client)
+
+                # Check if the ConfigMap already exists
+                response = api_instance.list_namespaced_config_map(
+                    namespace=self.namespace,
+                    field_selector=f"metadata.name={name}"
+                )
+                if response.items:
+                    api_instance.replace_namespaced_config_map(
+                        name=name,
+                        namespace=self.namespace,
+                        body=client.V1ConfigMap(
+                            api_version="v1",
+                            kind="ConfigMap",
+                            metadata=client.V1ObjectMeta(name=name),
+                            data=data
+                        )
+                    )
+                    logger.info(f"ConfigMap {name} already exists. Updating it")
+                    return
+
+                # Create a new ConfigMap if it doesn't exist
                 body = client.V1ConfigMap(
                     api_version="v1",
                     kind="ConfigMap",
@@ -28,7 +49,7 @@ class ConfigMap:
                 )
 
                 api_instance.create_namespaced_config_map(
-                    namespace=self.tenant,
+                    namespace=self.namespace,
                     body=body
                 )
         except ApiException as e:
@@ -37,7 +58,7 @@ class ConfigMap:
 
         logger.info(f"ConfigMap {name} created successfully")
 
-    def delete_config(self, name: str):
+    async def delete_config(self, name: str):
         """
         Delete the ConfigMap
         """
@@ -46,9 +67,19 @@ class ConfigMap:
         try:
             with client.ApiClient() as api_client:
                 api_instance = client.CoreV1Api(api_client)
+
+                # Check if the ConfigMap exists
+                response = api_instance.list_namespaced_config_map(
+                    namespace=self.namespace,
+                    field_selector=f"metadata.name={name}"
+                )
+                if not response.items:
+                    logger.info(f"ConfigMap {name} does not exist in namespace: {self.namespace}")
+                    return
+
                 api_instance.delete_namespaced_config_map(
                     name=name,
-                    namespace=self.tenant
+                    namespace=self.namespace
                 )
         except ApiException as e:
             logger.error(f"Error while deleting ConfigMap: {e}")
